refactor: move test stub out
